+++ Source/Math/Generic/FSGMath.cpp
@@ -1,6 +1,6 @@
 #include "Math/FSGMath.hpp"
 
-f64 FSGSqrt(const f32 &t_x) {
+f32 FSGSqrt(const f32 &t_x) {
     return sqrt(static_cast<f64>(t_x));
 }
